split schema to individual module
